LPS-96155 Not approved Content Page is published
