+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/exportimport/data/handler/StagedLayoutSetStagedModelDataHandler.java
@@ -37,7 +37,6 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.Image;
 import com.liferay.portal.kernel.model.Layout;
-import com.liferay.portal.kernel.model.LayoutRevision;
 import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.model.LayoutSetBranch;
 import com.liferay.portal.kernel.model.LayoutSetPrototype;
@@ -46,7 +45,6 @@ import com.liferay.portal.kernel.model.adapter.ModelAdapterUtil;
 import com.liferay.portal.kernel.service.GroupLocalService;
 import com.liferay.portal.kernel.service.ImageLocalService;
 import com.liferay.portal.kernel.service.LayoutLocalService;
-import com.liferay.portal.kernel.service.LayoutRevisionLocalService;
 import com.liferay.portal.kernel.service.LayoutSetBranchLocalService;
 import com.liferay.portal.kernel.service.LayoutSetLocalService;
 import com.liferay.portal.kernel.service.LayoutSetPrototypeLocalService;
@@ -61,7 +59,6 @@ import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.util.ThemeFactoryUtil;
 import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.kernel.workflow.WorkflowConstants;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.service.impl.LayoutLocalServiceHelper;
 import com.liferay.sites.kernel.util.Sites;
@@ -201,9 +198,7 @@ public class StagedLayoutSetStagedModelDataHandler
 							layoutUUID, stagingGroupID,
 							!layout.isPublicLayout());
 
-					if ((stagedLayout != null) &&
-						_isLayoutRevisionInReview(stagedLayout)) {
-
+					if (stagedLayout != null) {
 						continue;
 					}
 
@@ -886,23 +881,6 @@ public class StagedLayoutSetStagedModelDataHandler
 		}
 	}
 
-	private boolean _isLayoutRevisionInReview(Layout stagedLayout) {
-		List<LayoutRevision> layoutRevisions =
-			_layoutRevisionLocalService.getLayoutRevisions(
-				stagedLayout.getPlid());
-
-		Stream<LayoutRevision> layoutRevisionStream = layoutRevisions.stream();
-
-		if (layoutRevisionStream.anyMatch(
-				revision ->
-					revision.getStatus() == WorkflowConstants.STATUS_PENDING)) {
-
-			return true;
-		}
-
-		return false;
-	}
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		StagedLayoutSetStagedModelDataHandler.class);
 
@@ -924,9 +902,6 @@ public class StagedLayoutSetStagedModelDataHandler
 
 	@Reference
 	private LayoutLocalServiceHelper _layoutLocalServiceHelper;
-
-	@Reference
-	private LayoutRevisionLocalService _layoutRevisionLocalService;
 
 	@Reference
 	private LayoutSetBranchLocalService _layoutSetBranchLocalService;
